agent: rename eosagent to agent
