--- .github/workflows/golangci-lint.yml
@@ -1,31 +1,0 @@
-name: golangci-lint
-on:
-    push:
-      paths:
-        - "controlplane/**"
-        - "smartcontract/sdk/go/**"
-      branches: [ main ]
-    pull_request:
-      paths:
-        - "controlplane/**"
-        - "smartcontract/sdk/go/**"
-      branches: [ main ]
-
-permissions:
-    contents: read
-    checks: write
-
-jobs:
-    golangci:
-      name: lint
-      runs-on: ubuntu-latest
-      steps:
-        - uses: actions/checkout@v4
-        - uses: actions/setup-go@v5
-          with:
-            go-version: 1.24.x
-        - name: golangci-lint
-          uses: golangci/golangci-lint-action@v6
-          with:
-            version: v1.64.2
-            working-directory: controlplane
