+++ controlplane/agent/internal/agent/dzclient_test.go
@@ -1,4 +1,4 @@
-package eosagent
+package agent
 
 import (
 	"context"
